refactor: rename input/ to inputs/
