+++ web/src/components/FederatedConnectorSelector.tsx
@@ -5,7 +5,7 @@ import {
   federatedSourceToRegularSource,
 } from "@/lib/types";
 import { SourceIcon } from "@/components/SourceIcon";
-import { Label } from "@opal/layouts/input/components";
+import { Label } from "@opal/layouts/inputs/components";
 import { ErrorMessage } from "formik";
 import Text from "@/refresh-components/texts/Text";
 import InputTypeIn from "@/refresh-components/inputs/InputTypeIn";
