+++ web/src/sections/modals/NewTenantModal.tsx
@@ -10,7 +10,7 @@ import { useUser } from "@/providers/UserProvider";
 import { NewTenantInfo } from "@/lib/types";
 import { useRouter } from "next/navigation";
 import Text from "@/refresh-components/texts/Text";
-import { InputErrorText } from "@opal/layouts/input/components";
+import { InputErrorText } from "@opal/layouts/inputs/components";
 
 // App domain should not be hardcoded
 const APP_DOMAIN = process.env.NEXT_PUBLIC_APP_DOMAIN || "onyx.app";
