+++ web/src/sections/modals/llmConfig/AzureModal.tsx
@@ -3,7 +3,7 @@
 import { useSWRConfig } from "swr";
 import { useFormikContext } from "formik";
 import InputTypeInField from "@/refresh-components/form/InputTypeInField";
-import * as InputLayouts from "@opal/layouts/input/components";
+import * as InputLayouts from "@opal/layouts/inputs/components";
 import {
   LLMProviderFormProps,
   LLMProviderName,
