+++ web/src/sections/modals/llmConfig/OpenAICompatibleModal.tsx
@@ -3,7 +3,7 @@
 import { markdown } from "@opal/utils";
 import { useSWRConfig } from "swr";
 import { useFormikContext } from "formik";
-import * as InputLayouts from "@opal/layouts/input/components";
+import * as InputLayouts from "@opal/layouts/inputs/components";
 import {
   LLMProviderFormProps,
   LLMProviderName,
