+++ web/src/sections/modals/llmConfig/OpenRouterModal.tsx
@@ -2,7 +2,7 @@
 
 import { useSWRConfig } from "swr";
 import { useFormikContext } from "formik";
-import * as InputLayouts from "@opal/layouts/input/components";
+import * as InputLayouts from "@opal/layouts/inputs/components";
 import {
   LLMProviderFormProps,
   LLMProviderName,
